refactor: abstract agg committee msg validation for fork to helper
